Install examples in example directory instead of bin. Fix example02 run number and event number type.
git-svn-id: https://phys12svn.jlab.org/repos/JANA@2380 c5ed4466-e916-0410-8347-b3263e9c103d

+++ src/examples/example02/JEventProcessor_MyProcessor.cc
@@ -42,7 +42,7 @@ jerror_t JEventProcessor_MyProcessor::init(void)
 //------------------
 // brun
 //------------------
-jerror_t JEventProcessor_MyProcessor::brun(JEventLoop *eventLoop, int runnumber)
+jerror_t JEventProcessor_MyProcessor::brun(JEventLoop *eventLoop, int32_t runnumber)
 {
 	// This is called whenever the run number changes
 	return NOERROR;
@@ -51,7 +51,7 @@ jerror_t JEventProcessor_MyProcessor::brun(JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t JEventProcessor_MyProcessor::evnt(JEventLoop *loop, int eventnumber)
+jerror_t JEventProcessor_MyProcessor::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 	// This is called for every event. Use of common resources like writing
 	// to a file or filling a histogram should be mutex protected. Using
